requirements.txt
google-generativeai
pandas
numpy
pillow
opencv-python
gradio
python-dotenv
